Change to depend on matplotlib-base
We don't *require* the GUI, so don't depend on matplotlib, just
matplotlib-base. This change to is allow metpy to install alongside
sharppy on linux. Right now, matplotlib (depends on pyqt) cannot install
next to sharppy (depends on pyside2). This will probably be fine once
qt 5.12 builds of pyqt are available.